# Use parent to determine whether an expression is target of an invocation