--- a/runtime/sema/check_for.go
+++ b/runtime/sema/check_for.go
@@ -41,7 +41,7 @@ func (checker *Checker) VisitForStatement(statement *ast.ForStatement) (_ struct
 		}
 	}
 
-	valueType := checker.VisitExpression(valueExpression, expectedType)
+	valueType := checker.VisitExpression(valueExpression, statement, expectedType)
 
 	// Only get the element type if the array is not a resource array.
 	// Otherwise, in addition to the `UnsupportedResourceForLoopError`,