--- a/runtime/sema/check_remove_statement.go
+++ b/runtime/sema/check_remove_statement.go
@@ -32,7 +32,7 @@ func (checker *Checker) VisitRemoveStatement(statement *ast.RemoveStatement) (_ 
 	}
 
 	nominalType := checker.convertNominalType(statement.Attachment)
-	base := checker.VisitExpression(statement.Value, nil)
+	base := checker.VisitExpression(statement.Value, statement, nil)
 	checker.checkUnusedExpressionResourceLoss(base, statement.Value)
 
 	if nominalType == InvalidType {