Port https://github.com/dapperlabs/cadence-internal/pull/220

Closes https://github.com/dapperlabs/cadence-internal/issues/213

Also lays the foundation for https://github.com/onflow/cadence/pull/3341

## Description

Concept is same as https://github.com/dapperlabs/cadence-internal/pull/146, where parent is tracked to determine whether the current expression is an invoked-expression.

However, unlike in https://github.com/dapperlabs/cadence-internal/pull/146, only record the parent temporarily as the checker visit the expression, and then reset it. Doesn't store the parent-child mapping in the elaboration. This way, there is no performance impact:
```
                                                 │   before    │               after               │
                                                 │   sec/op    │   sec/op     vs base              │
CheckContractInterfaceFungibleTokenConformance-8   65.28µ ± 0%   65.58µ ± 0%  +0.46% (p=0.009 n=6)

                                                 │    before    │               after                │
                                                 │     B/op     │     B/op      vs base              │
CheckContractInterfaceFungibleTokenConformance-8   57.77Ki ± 0%   57.78Ki ± 0%  +0.03% (p=0.002 n=6)

                                                 │   before    │             after              │
                                                 │  allocs/op  │  allocs/op   vs base           │
CheckContractInterfaceFungibleTokenConformance-8   1.022k ± 0%   1.022k ± 0%  ~ (p=1.000 n=6) ¹
```

______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels