@SupunS 2024-05-14
> I wonder if we can also refactor other existing flags, like inAssignment (in a follow-up PR)

Yeah, would be great if we can remove these other flags as well. I'll create an issue for this.